+++ dynamo3/connection.py
@@ -1,19 +1,18 @@
 """ Connection class for DynamoDB """
-import re
 import time
 import warnings
-import inspect
 
 import botocore.session
 import six
 from botocore.exceptions import ClientError
 
 from .batch import BatchWriter, encode_query_kwargs
-from .constants import NONE, COUNT
+from .constants import NONE, COUNT, INDEXES, READ_COMMANDS
 from .exception import translate_exception, DynamoDBError, ThroughputException
 from .fields import Throughput, Table
-from .result import ResultSet, GetResultSet, Result, Count
+from .result import ResultSet, GetResultSet, Result, Count, ConsumedCapacity
 from .types import Dynamizer, is_null
+
 
 
 def build_expected(dynamizer, expected):
@@ -61,6 +60,9 @@ class DynamoDBConnection(object):
     request_retries : int
         Number of times to retry an API call if the throughput is exceeded
         (default 10)
+    default_return_capacity : bool
+        If true, all relevant calls will default to fetching the
+        ConsumedCapacity
 
     """
 
@@ -68,6 +70,13 @@ class DynamoDBConnection(object):
         self.client = client
         self.dynamizer = dynamizer
         self.request_retries = 10
+        self.default_return_capacity = False
+        self.last_consumed_capacity = None
+        self._hooks = {
+            'precall': [],
+            'postcall': [],
+            'capacity': [],
+        }
 
     @property
     def host(self):
@@ -206,7 +215,10 @@ class DynamoDBConnection(object):
         data : dict
 
         """
+        for hook in self._hooks['precall']:
+            hook(self, command, kwargs)
         op = getattr(self.client, command)
+        self.last_consumed_capacity = None
         attempt = 0
         while True:
             try:
@@ -221,12 +233,43 @@ class DynamoDBConnection(object):
                     self.exponential_sleep(attempt)
                 else:
                     exc.re_raise()
+        for hook in self._hooks['postcall']:
+            hook(self, command, kwargs, data)
+        if 'ConsumedCapacity' in data:
+            is_read = command in READ_COMMANDS
+            consumed = data['ConsumedCapacity']
+            if isinstance(consumed, list):
+                self.last_consumed_capacity = \
+                    [ConsumedCapacity.from_response(cap, is_read) for cap in
+                     consumed]
+                all_caps = self.last_consumed_capacity
+            else:
+                self.last_consumed_capacity = \
+                    ConsumedCapacity.from_response(consumed, is_read)
+                all_caps = [self.last_consumed_capacity]
+            for hook in self._hooks['capacity']:
+                for cap in all_caps:
+                    hook(self, command, kwargs, data, cap)
         return data
 
     def exponential_sleep(self, attempt):
         """ Sleep with exponential backoff """
         if attempt > 1:
             time.sleep(0.1 * 2 ** attempt)
+
+    def subscribe(self, event, hook):
+        self._hooks[event].append(hook)
+
+    def unsubscribe(self, event, hook):
+        self._hooks[event].remove(hook)
+
+    def _default_capacity(self, value):
+        """ Get the value for ReturnConsumedCapacity from provided value """
+        if value is not None:
+            return value
+        if self.default_return_capacity:
+            return INDEXES
+        return NONE
 
     def list_tables(self, limit=100):
         """
@@ -270,7 +313,7 @@ class DynamoDBConnection(object):
                 raise
 
     def create_table(self, tablename, hash_key, range_key=None,
-                     indexes=None, global_indexes=None, throughput=None, dry_run=False):
+                     indexes=None, global_indexes=None, throughput=None):
         """
         Create a table
 
@@ -322,11 +365,9 @@ class DynamoDBConnection(object):
 
         kwargs['AttributeDefinitions'] = [attr.definition() for attr in
                                           all_attrs]
-        if dry_run:
-            return kwargs
         return self.call('create_table', **kwargs)
 
-    def delete_table(self, tablename, dry_run=False):
+    def delete_table(self, tablename):
         """
         Delete a table
 
@@ -342,8 +383,6 @@ class DynamoDBConnection(object):
 
         """
         try:
-            if dry_run:
-                return {'TableName': tablename}
             self.call('delete_table', TableName=tablename)
             return True
         except DynamoDBError as e:
@@ -353,8 +392,7 @@ class DynamoDBConnection(object):
                 raise
 
     def put_item(self, tablename, item, expected=None, returns=NONE,
-                 return_capacity=NONE, expect_or=False, dry_run=False,
-                 **kwargs):
+                 return_capacity=None, expect_or=False, **kwargs):
         """
         Store an item, overwriting existing data
 
@@ -393,20 +431,19 @@ class DynamoDBConnection(object):
             if len(keywords['Expected']) > 1:
                 keywords['ConditionalOperator'] = 'OR' if expect_or else 'AND'
         elif expected is not None:
-            warnings.warn("Using deprecated argument 'expected' for put_item. "
-                          "Use kwargs instead.")
             keywords['Expected'] = build_expected(self.dynamizer, expected)
+        keywords['ReturnConsumedCapacity'] = \
+            self._default_capacity(return_capacity)
         item = self.dynamizer.encode_keys(item)
         ret = self.call('put_item', TableName=tablename, Item=item,
-                        ReturnValues=returns,
-                        ReturnConsumedCapacity=return_capacity, **keywords)
+                        ReturnValues=returns, **keywords)
         if ret:
-            return Result(self.dynamizer, ret, 'Attributes', is_read=False)
+            return Result(self.dynamizer, ret, 'Attributes',
+                          self.last_consumed_capacity)
 
     def put_item2(self, tablename, item, expr_values=None, alias=None,
-                  condition=None, returns=NONE, return_capacity=NONE,
-                  return_item_collection_metrics=NONE, dry_run=False,
-                  **kwargs):
+                  condition=None, returns=NONE, return_capacity=None,
+                  return_item_collection_metrics=NONE, **kwargs):
         """
         Put a new item into a table
 
@@ -445,7 +482,7 @@ class DynamoDBConnection(object):
             'TableName': tablename,
             'Item': self.dynamizer.encode_keys(item),
             'ReturnValues': returns,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
             'ReturnItemCollectionMetrics': return_item_collection_metrics,
         }
         values = build_expression_values(self.dynamizer, expr_values, kwargs)
@@ -455,14 +492,13 @@ class DynamoDBConnection(object):
             keywords['ExpressionAttributeNames'] = alias
         if condition:
             keywords['ConditionExpression'] = condition
-        if dry_run:
-            return keywords
         result = self.call('put_item', **keywords)
         if result:
-            return Result(self.dynamizer, result, 'Attributes', is_read=False)
+            return Result(self.dynamizer, result, 'Attributes',
+                          self.last_consumed_capacity)
 
     def get_item(self, tablename, key, attributes=None, consistent=False,
-                 return_capacity=NONE):
+                 return_capacity=None):
         """
         Fetch a single item from a table
 
@@ -490,15 +526,16 @@ class DynamoDBConnection(object):
             'TableName': tablename,
             'Key': self.dynamizer.encode_keys(key),
             'ConsistentRead': consistent,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
         }
         if attributes is not None:
             kwargs['AttributesToGet'] = attributes
         data = self.call('get_item', **kwargs)
-        return Result(self.dynamizer, data, 'Item', is_read=True)
+        return Result(self.dynamizer, data, 'Item',
+                      self.last_consumed_capacity)
 
     def get_item2(self, tablename, key, attributes=None, alias=None,
-                  consistent=False, return_capacity=NONE, dry_run=False):
+                  consistent=False, return_capacity=None):
         """
         Fetch a single item from a table
 
@@ -526,7 +563,7 @@ class DynamoDBConnection(object):
             'TableName': tablename,
             'Key': self.dynamizer.encode_keys(key),
             'ConsistentRead': consistent,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
         }
         if attributes is not None:
             if not isinstance(attributes, six.string_types):
@@ -534,13 +571,12 @@ class DynamoDBConnection(object):
             kwargs['ProjectionExpression'] = attributes
         if alias:
             kwargs['ExpressionAttributeNames'] = alias
-        if dry_run:
-            return kwargs
         data = self.call('get_item', **kwargs)
-        return Result(self.dynamizer, data, 'Item', is_read=True)
+        return Result(self.dynamizer, data, 'Item',
+                      self.last_consumed_capacity)
 
     def delete_item(self, tablename, key, expected=None, returns=NONE,
-                    return_capacity=NONE, expect_or=False, **kwargs):
+                    return_capacity=None, expect_or=False, **kwargs):
         """
         Delete an item
 
@@ -574,25 +610,24 @@ class DynamoDBConnection(object):
 
         """
         key = self.dynamizer.encode_keys(key)
-        keywords = {}
+        keywords = {
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
+        }
         if kwargs:
             keywords['Expected'] = encode_query_kwargs(self.dynamizer, kwargs)
             if len(keywords['Expected']) > 1:
                 keywords['ConditionalOperator'] = 'OR' if expect_or else 'AND'
         elif expected is not None:
-            warnings.warn("Using deprecated argument 'expected' for "
-                          "delete_item. Use kwargs instead.")
             keywords['Expected'] = build_expected(self.dynamizer, expected)
         ret = self.call('delete_item', TableName=tablename, Key=key,
-                        ReturnValues=returns,
-                        ReturnConsumedCapacity=return_capacity,
-                        **keywords)
+                        ReturnValues=returns, **keywords)
         if ret:
-            return Result(self.dynamizer, ret, 'Attributes', is_read=False)
+            return Result(self.dynamizer, ret, 'Attributes',
+                          self.last_consumed_capacity)
 
     def delete_item2(self, tablename, key, expr_values=None, alias=None,
-                     condition=None, returns=NONE, return_capacity=NONE,
-                     return_item_collection_metrics=NONE, dry_run=False, **kwargs):
+                     condition=None, returns=NONE, return_capacity=None,
+                     return_item_collection_metrics=NONE, **kwargs):
         """
         Delete an item from a table
 
@@ -632,7 +667,7 @@ class DynamoDBConnection(object):
             'TableName': tablename,
             'Key': self.dynamizer.encode_keys(key),
             'ReturnValues': returns,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
             'ReturnItemCollectionMetrics': return_item_collection_metrics,
         }
         values = build_expression_values(self.dynamizer, expr_values, kwargs)
@@ -642,14 +677,13 @@ class DynamoDBConnection(object):
             keywords['ExpressionAttributeNames'] = alias
         if condition:
             keywords['ConditionExpression'] = condition
-        if dry_run:
-            return keywords
         result = self.call('delete_item', **keywords)
         if result:
-            return Result(self.dynamizer, result, 'Attributes', is_read=False)
+            return Result(self.dynamizer, result, 'Attributes',
+                          self.last_consumed_capacity)
 
-    def batch_write(self, tablename, return_capacity=NONE,
-                    return_item_collection_metrics=NONE, dry_run=False):
+    def batch_write(self, tablename, return_capacity=None,
+                    return_item_collection_metrics=NONE):
         """
         Perform a batch write on a table
 
@@ -657,6 +691,13 @@ class DynamoDBConnection(object):
         ----------
         tablename : str
             Name of the table to write to
+        return_capacity : {NONE, INDEXES, TOTAL}, optional
+            INDEXES will return the consumed capacity for indexes, TOTAL will
+            return the consumed capacity for the table and the indexes.
+            (default NONE)
+        return_item_collection_metrics : (NONE, SIZE), optional
+            SIZE will return statistics about item collections that were
+            modified.
 
         Examples
         --------
@@ -667,12 +708,12 @@ class DynamoDBConnection(object):
                 batch.delete({'id': 'oldid'})
 
         """
+        return_capacity = self._default_capacity(return_capacity)
         return BatchWriter(self, tablename, return_capacity=return_capacity,
-                           return_item_collection_metrics=return_item_collection_metrics,
-                           dry_run=dry_run)
+                           return_item_collection_metrics=return_item_collection_metrics)
 
     def batch_get(self, tablename, keys, attributes=None, consistent=False,
-                  return_capacity=NONE, dry_run=False):
+                  return_capacity=None):
         """
         Perform a batch get of many items in a table
 
@@ -694,15 +735,14 @@ class DynamoDBConnection(object):
 
         """
         keys = [self.dynamizer.encode_keys(k) for k in keys]
+        return_capacity = self._default_capacity(return_capacity)
         ret = GetResultSet(self, tablename, keys,
                            consistent=consistent, attributes=attributes,
                            return_capacity=return_capacity)
-        if dry_run:
-            return ret.build_kwargs()
         return ret
 
     def update_item(self, tablename, key, updates, returns=NONE,
-                    return_capacity=NONE, expect_or=False, **kwargs):
+                    return_capacity=None, expect_or=False, **kwargs):
         """
         Update a single item in a table
 
@@ -745,7 +785,9 @@ class DynamoDBConnection(object):
         key = self.dynamizer.encode_keys(key)
         attr_updates = {}
         expected = {}
-        keywords = {}
+        keywords = {
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
+        }
         for update in updates:
             attr_updates.update(update.attrs(self.dynamizer))
             expected.update(update.expected(self.dynamizer))
@@ -764,14 +806,14 @@ class DynamoDBConnection(object):
         result = self.call('update_item', TableName=tablename, Key=key,
                            AttributeUpdates=attr_updates,
                            ReturnValues=returns,
-                           ReturnConsumedCapacity=return_capacity,
                            **keywords)
         if result:
-            return Result(self.dynamizer, result, 'Attributes', is_read=False)
+            return Result(self.dynamizer, result, 'Attributes',
+                          self.last_consumed_capacity)
 
     def update_item2(self, tablename, key, expression, expr_values=None, alias=None,
-                     condition=None, returns=NONE, return_capacity=NONE,
-                     return_item_collection_metrics=NONE, dry_run=False, **kwargs):
+                     condition=None, returns=NONE, return_capacity=None,
+                     return_item_collection_metrics=NONE, **kwargs):
         """
         Update a single item in a table
 
@@ -814,7 +856,7 @@ class DynamoDBConnection(object):
             'Key': self.dynamizer.encode_keys(key),
             'UpdateExpression': expression,
             'ReturnValues': returns,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
             'ReturnItemCollectionMetrics': return_item_collection_metrics,
         }
         values = build_expression_values(self.dynamizer, expr_values, kwargs)
@@ -824,14 +866,13 @@ class DynamoDBConnection(object):
             keywords['ExpressionAttributeNames'] = alias
         if condition:
             keywords['ConditionExpression'] = condition
-        if dry_run:
-            return keywords
         result = self.call('update_item', **keywords)
         if result:
-            return Result(self.dynamizer, result, 'Attributes', is_read=False)
+            return Result(self.dynamizer, result, 'Attributes',
+                          self.last_consumed_capacity)
 
     def scan(self, tablename, attributes=None, count=False, limit=None,
-             return_capacity=NONE, filter_or=False, **kwargs):
+             return_capacity=None, filter_or=False, **kwargs):
         """
         Perform a full-table scan
 
@@ -873,7 +914,7 @@ class DynamoDBConnection(object):
         """
         keywords = {
             'TableName': tablename,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
         }
         if attributes is not None:
             keywords['AttributesToGet'] = attributes
@@ -892,8 +933,8 @@ class DynamoDBConnection(object):
 
     def scan2(self, tablename, expr_values=None, alias=None, attributes=None,
               consistent=False, select=None, index=None, limit=None,
-              return_capacity=NONE, filter=False, segment=None,
-              total_segments=None, dry_run=False, **kwargs):
+              return_capacity=None, filter=False, segment=None,
+              total_segments=None, **kwargs):
         """
         Perform a full-table scan
 
@@ -947,7 +988,7 @@ class DynamoDBConnection(object):
         """
         keywords = {
             'TableName': tablename,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
             'ConsistentRead': consistent,
         }
         values = build_expression_values(self.dynamizer, expr_values, kwargs)
@@ -972,15 +1013,14 @@ class DynamoDBConnection(object):
         if total_segments is not None:
             keywords['TotalSegments'] = total_segments
 
-        if dry_run:
-            return keywords
-        elif select == COUNT:
-            return Count.from_response(self.call('scan', **keywords))
+        if select == COUNT:
+            return Count.from_response(self.call('scan', **keywords),
+                                       self.last_consumed_capacity)
         else:
             return ResultSet(self, 'Items', 'scan', **keywords)
 
     def query(self, tablename, attributes=None, consistent=False, count=False,
-              index=None, limit=None, desc=False, return_capacity=NONE,
+              index=None, limit=None, desc=False, return_capacity=None,
               filter=None, filter_or=False, **kwargs):
         """
         Perform an index query on a table
@@ -1033,7 +1073,7 @@ class DynamoDBConnection(object):
         """
         keywords = {
             'TableName': tablename,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
             'ConsistentRead': consistent,
         }
         if attributes is not None:
@@ -1054,14 +1094,15 @@ class DynamoDBConnection(object):
                                                         kwargs)
         if count:
             keywords['Select'] = COUNT
-            return Count.from_response(self.call('query', **keywords))
+            return Count.from_response(self.call('query', **keywords),
+                                       self.last_consumed_capacity)
         else:
             return ResultSet(self, 'Items', 'query', **keywords)
 
     def query2(self, tablename, key_condition_expr, expr_values=None,
                alias=None, attributes=None, consistent=False, select=None,
-               index=None, limit=None, desc=False, return_capacity=NONE,
-               filter=None, dry_run=False, **kwargs):
+               index=None, limit=None, desc=False, return_capacity=None,
+               filter=None, **kwargs):
         """
         Perform an index query on a table
 
@@ -1114,7 +1155,7 @@ class DynamoDBConnection(object):
         values = build_expression_values(self.dynamizer, expr_values, kwargs)
         keywords = {
             'TableName': tablename,
-            'ReturnConsumedCapacity': return_capacity,
+            'ReturnConsumedCapacity': self._default_capacity(return_capacity),
             'ConsistentRead': consistent,
             'KeyConditionExpression': key_condition_expr,
             'ExpressionAttributeValues': values,
@@ -1135,15 +1176,14 @@ class DynamoDBConnection(object):
         if filter:
             keywords['FilterExpression'] = filter
 
-        if dry_run:
-            return keywords
-        elif select == COUNT:
-            return Count.from_response(self.call('query', **keywords))
+        if select == COUNT:
+            return Count.from_response(self.call('query', **keywords),
+                                       self.last_consumed_capacity)
         else:
             return ResultSet(self, 'Items', 'query', **keywords)
 
     def update_table(self, tablename, throughput=None, global_indexes=None,
-                     index_updates=None, dry_run=False):
+                     index_updates=None):
         """
         Update the throughput of a table and/or global indexes
 
@@ -1185,6 +1225,4 @@ class DynamoDBConnection(object):
         if all_attrs:
             attr_definitions = [attr.definition() for attr in all_attrs]
             kwargs['AttributeDefinitions'] = attr_definitions
-        if dry_run:
-            return kwargs
         return self.call('update_table', **kwargs)